refactor(phase-19)(pr-4a): trade_clue 디렉터리 승격
cluedist/trade_clue.go (532줄) → cluedist/trade_clue/ 디렉터리 분할.

파일 구조:
- module.go — 구조체, NewTradeClueModule, Init, Cleanup, engine.Module assertion, init()
- config.go — TradeClueConfig, TradeProposal, ShowSession, Schema
- handlers.go — HandleMessage + trade_propose/accept/decline + show_request/accept/decline
- reactor.go — ReactTo (ALLOW_EXCHANGE), SupportedActions
- state.go — BuildState, BuildStateFor (PlayerAware), SaveState, RestoreState

cluedist/register.go 신설: trade_clue 하위 package blank-import.

Factory 서명, engine.Register 방식, PlayerAware/PublicStateMarker 모두 불변.
F-go-3 해소 (500+ 파일 4→3).
